requirements.txt
autopep8==1.4
certifi==2018.8.24
chardet==3.0.4
Django==2.1
django-debug-toolbar==1.10.1
django-widget-tweaks==1.4.3
idna==2.7
mysqlclient==1.3.13
pycodestyle==2.4.0
pytz==2018.5
requests==2.19.1
sqlparse==0.2.4
urllib3==1.23
